requirements.txt
google-genai
python-dotenv
streamlit
typing-extensions
numpy
pandas
requests
scikit-learn>=1.3.0
